Fixed ruler alignment on MacOS for fonts with fractional metrics.
Task-number: QTCREATORBUG-1065
(cherry picked from commit fbb8f39945dd3991353a33ff783aae5ae2e9f8e8)

+++ src/plugins/texteditor/basetexteditor.cpp
@@ -2987,13 +2987,13 @@ void BaseTextEditor::extraAreaPaintEvent(QPaintEvent *e)
 
     QTextBlock block = firstVisibleBlock();
     int blockNumber = block.blockNumber();
-    int top = (int) blockBoundingGeometry(block).translated(contentOffset()).top();
-    int bottom = top;
+    qreal top = blockBoundingGeometry(block).translated(contentOffset()).top();
+    qreal bottom = top;
 
     while (block.isValid() && top <= e->rect().bottom()) {
 
         top = bottom;
-        bottom = top + (int)blockBoundingRect(block).height();
+        bottom = top + blockBoundingRect(block).height();
         QTextBlock nextBlock = block.next();
 
         QTextBlock nextVisibleBlock = nextBlock;
